% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{BronCell.plot}
\alias{BronCell.plot}
\title{plot cell proporions for each bronchial cell types}
\usage{
BronCell.plot(est_prop)
}
\arguments{
\item{est_prop}{cell proportion matrix, estimated by funtion BronCell.prop}
}
\description{
plot cell proporions for each bronchial cell types
}
