% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{BronCell.prop}
\alias{BronCell.prop}
\title{main function of this package, to calculate cell proportion}
\usage{
BronCell.prop(bulk.eset, method = "MuSiC")
}
\arguments{
\item{bulk.eset}{ExpressionSet for bulk data, raw counts as the input}

\item{method}{Three deconvolution methods are provided: MuSiC, NNLS and bseqsc (based on CIBERSORT), MuSiC is the defual method}
}
\value{
est.prop, A cell proportion matrix, each row represent a sample, each col is a cell type
}
\description{
main function of this package, to calculate cell proportion
}
