% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{music_prop_bron}
\alias{music_prop_bron}
\title{modified function of "music_prop" from MuSiC package using the signature matrix as the input
original code is on: https://github.com/xuranw/MuSiC}
\usage{
music_prop_bron(
  bulk.eset,
  sc.basis,
  clusters,
  samples,
  select.ct = NULL,
  cell_size = NULL,
  ct.cov = FALSE,
  verbose = TRUE,
  iter.max = 1000,
  nu = 1e-04,
  eps = 0.01,
  centered = FALSE,
  normalize = FALSE,
  ...
)
}
\arguments{
\item{bulk.eset}{ExpressionSet for bulk data}

\item{sc.basis}{Gene signature matrix generated from bronchial biopsy scRNAseq data, using function music_basis from MuSiC package}

\item{clusters}{character, the phenoData of single cell dataset used as clusters;}

\item{samples}{character,the phenoData of single cell dataset used as samples;}

\item{select.ct}{vector of cell types, default as NULL. If NULL, then use all cell types provided by single cell dataset;}

\item{cell_size}{data.frame of cell sizes. 1st column contains the names of cell types, 2nd column has the cell sizes per cell type. Default as NULL. If NULL, then estimate cell size from data;}

\item{ct.cov}{logical. If TRUE, use the covariance across cell types;}

\item{verbose}{logical, default as TRUE.}

\item{iter.max}{numeric, maximum iteration number}

\item{nu}{regulation parameter, take care of weight when taking recipical}

\item{eps}{Thredshold of convergence}

\item{centered}{logic, substract avg of Y and D}

\item{normalize}{logic, divide Y and D by their standard deviation}
}
\value{
a list with elements:
   * Estimates of MuSiC
   * Estimates of NNLS
   * Weight of MuSiC
   * r.squared of MuSiC
   * Variance of MuSiC estimates
}
\description{
This function is to calculate the MuSiC deconvolution proportions
}
