% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc.basis}
\alias{sc.basis}
\title{gene signature matrix used for predicting cell proportion from bulk RNAseq data
generated from MuSiC fuction 'musci_basis'}
\format{
a music basis object with signature matrix before and after scaling, names are coded in gene symbol
}
\usage{
sc.basis
}
\description{
gene signature matrix used for predicting cell proportion from bulk RNAseq data
generated from MuSiC fuction 'musci_basis'
}
\keyword{datasets}
